Extract a namespace for Screen
